Ran into same old controller_manager inability to load robot_description outside of / https://github.com/ros-simulation/gazebo_ros_pkgs/issues/112 and http://answers.ros.org/question/198929/running-controller_manager-spawner-with-mybotrobot_description/, which is apparently fixed, but not in released packages?  Need to follow up with that.
Made the pop can bigger but now it doesn't spawn in the fridge, maybe should just make own cylinder object with a xacro and all so it can be controlled better, just make it solid red.

Was running into problems with two joint_state_publishers (though the namespace solution above would solve it, if it actually is working?), so disabled the fridge joint_state_publisher.

The friction coefficients work okay but not great, still need to drive very slowly.